Crate bls12_381

bls12_381

This crate provides an implementation of the BLS12-381 pairing-friendly elliptic curve construction.

This implementation has not been reviewed or audited. Use at your own risk.

This implementation targets Rust 1.36 or later.

This implementation does not require the Rust standard library.

All operations are constant time unless explicitly noted.

Modules

notes

Notes about how the BLS12-381 elliptic curve is designed, specified and implemented by this library.

Structs

Scalar

Represents an element of the scalar field $\mathbb{F}_q$ of the BLS12-381 elliptic curve construction.